adding move cursor to policy field key; deleting policy field key in the dryrun request; fixing z-index problem for boolean policy criteria
