+++ ui/apps/platform/src/Containers/Policies/PatternFly/Wizard/Step3/BooleanPolicyLogicSection.tsx
@@ -6,8 +6,8 @@ import {
     policyConfigurationDescriptor,
     networkDetectionDescriptor,
     auditLogDescriptor,
-    Descriptor,
 } from 'Containers/Policies/Wizard/Form/descriptors';
+import { Descriptor } from 'Containers/Policies/Wizard/Form/descriptor';
 import { Policy } from 'types/policy.proto';
 import PolicySection from './PolicySection';
