+++ ui/apps/platform/src/Containers/Policies/PatternFly/Wizard/Step3/PolicyCriteriaFieldValue.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Flex, Button } from '@patternfly/react-core';
 import { TimesIcon } from '@patternfly/react-icons';
 
-import { Descriptor } from 'Containers/Policies/Wizard/Form/descriptors';
+import { Descriptor } from 'Containers/Policies/Wizard/Form/descriptor';
 import PolicyCriteriaFieldInput from './PolicyCriteriaFieldInput';
 
 type FieldValueProps = {
