+++ ui/apps/platform/src/Containers/Policies/PatternFly/Wizard/Step3/PolicyCriteriaKey.css
@@ -1,5 +1,6 @@
 .policy-criteria-key {
     border: var(--pf-global--BorderColor--100) 1px solid;
+    cursor: move;
 }
 
 .draggable-grip {
@@ -8,7 +9,6 @@
     font-family: sans-serif;
     line-height: 4px;
     margin-right: 0.3em;
-    cursor: move;
 }
 
 .policy-criteria-key .draggable-grip::after {
