+++ ui/apps/platform/src/Containers/Policies/PatternFly/Wizard/Step3/PolicySection.css
@@ -27,6 +27,6 @@
         background-color: var(--pf-global--BackgroundColor--100);
         position: sticky;
         top: 0;
-        z-index: 100;
+        z-index: 101;
     }
 }
